added Pillow image verification
